Refactored code into wrapper , handler and reproter modules.
